% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampArea.R
\name{sampArea}
\alias{sampArea}
\title{Create a sampling area}
\usage{
sampArea(region, edgeBuffer = NULL, roads, roadBufferMin = 100,
  roadBufferMax = 2100, exclude = NULL, utmZone = 13)
}
\arguments{
\item{region}{the larger area to consider (a \code{SpatialPolygonsDataFrame})}

\item{edgeBuffer}{an optional buffer to exclude edge effects}

\item{roads}{a \code{SpatialLinesDataFrame} of access roads}

\item{roadBufferMin}{the buffer to exclude around roads (in meters)}

\item{roadBufferMax}{the max distance from roads within which points will be created 
(in meters)}

\item{exclude}{an optional additional area to exclude as a \code{Spatial}* object}

\item{utmZone}{the UTM zone to use for reprojection to UTM}
}
\value{
A \code{SpatialPointsDataFrame} containing randomized sampling points
}
\description{
Function to create an area for generating sampling points within
}
\details{
Regardless of the CRS of the input polygons, thus function internally converts
to and returns a polygon in a UTM projection, which is important for distance
calculations.
}
\author{
Andy Rominger <ajrominger@gmail.com>
}
