% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampPoints.R
\name{sampPoints}
\alias{sampPoints}
\title{Create randomized sampling points}
\usage{
sampPoints(n, area, dmin, name = "pnt", seed = 123)
}
\arguments{
\item{n}{number of points desired}

\item{area}{the spatial polygons within which points will be made}

\item{dmin}{minimum distance between points}

\item{name}{name to be applied to points in the form <name>_01, <name>_02, ...}

\item{seed}{random seed}
}
\value{
A \code{SpatialPointsDataFrame} containing randomized sampling points including
a column called \code{name} that is neccesary for output to some formats
}
\description{
Function to create randomized sampling points within spatial polygons, 
potentially with additional constraints such as roads
}
\details{
This function uses \code{spatstat::rSSI} to generate points. Note that if 
\code{spatstat::rSSI} cannot return the desired number of points it will issue a 
warning but not an error. If the desired number of points cannot be generated this 
is likely because \code{dmin} is too big for the \code{area} and one or both should be 
adjusted.
}
\author{
Andy Rominger <ajrominger@gmail.com>
}
