% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeMultiOGR.R
\name{writeMultiOGR}
\alias{writeMultiOGR}
\title{Wrapper around \code{rgdal::writeOGR} to succinctly write multiple files}
\usage{
writeMultiOGR(obj, dsn, layer, driver, ...)
}
\arguments{
\item{obj}{the object to write}

\item{dsn}{the path to the file to be written}

\item{layer}{the layer name}

\item{driver}{a character vector giving the name(s) of drivers to use}

\item{...}{other arguments passed to \code{rgdal::writeOGR}}
}
\value{
See \code{rgdal::writeOGR}
}
\description{
For every driver name in \code{driver}, thus function makes an 
appropriate call to \code{rgdal::writeOGR}
}
\details{
See \code{rgdal::writeOGR}
}
\author{
Andy Rominger <ajrominger@gmail.com>
}
