% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadCaatingaSpatial.R
\name{loadCaatingaSpatial}
\alias{loadCaatingaSpatial}
\title{Load the spatial data you'll need}
\usage{
loadCaatingaSpatial(message = "Here ya go...")
}
\arguments{
\item{message}{a nice thing to say to yourself}
}
\value{
Creats the following objects:
\describe{
  \item{\code{forest}}{SOSMA fragments as a \code{SpatialPolygonsDataFrame}}
  \item{\code{lcover}}{SOSMA landcover classes as a \code{raster}}
  \item{\code{lcoverke}}{key to SOSMA landcover as a \code{data.frame}}
  \item{\code{regions}}{a \code{SpatialPolygonsDataFrame} that delineates the different regions}
  \item{\code{roads}}{a \code{SpatialLinesDataFrame} of the roads}
}
}
\description{
A helper function to load raw spatial data
}
\details{
Runs \code{source} on a system file that contains code to load other system 
files containing needed spatial data
}
\author{
Andy Rominger <ajrominger@gmail.com>
}
